refactor(headless): split RSI loop replay tests
