+++ packages/headless/src/__tests__/prompt-optimization-loop.test.ts
@@ -1,1099 +1,11 @@
 import assert from 'node:assert/strict';
-import { execFile } from 'node:child_process';
-import { mkdir, mkdtemp, readFile, rm, writeFile } from 'node:fs/promises';
-import { tmpdir } from 'node:os';
+import { readFile } from 'node:fs/promises';
 import { join } from 'node:path';
-import { promisify } from 'node:util';
 import { describe, test } from 'node:test';
-import { hashSystemPrompt, readFixedPromptWal, type HarborTaskRunInput, type HarborTaskRunOutput } from '../fixed-prompt-controller.js';
-import { createCliPromptCandidateGit, type MetaAgent, type MetaAgentPromptInput } from '../prompt-candidate-loop.js';
-import { runPromptOptimizationLoop, type PromptOptimizationLoopInput } from '../prompt-optimization-loop.js';
-import type { Config } from '../contracts.js';
-import { tokenSummary } from './helpers/cell-output-fixtures.js';
-
-const execFileAsync = promisify(execFile);
-
-const CONFIG: Config = { id: 'cfg', backend: 'fake', llmConnectionSlug: 'deepseek' };
-const COST_PER_TASK = 0.02;
+import { readFixedPromptWal } from '../fixed-prompt-controller.js';
+import { execFileAsync, evidenceRefsFor, makeTasks, runLoop, taskIndex, withHarness, type MetaAgentPromptInput } from './helpers/prompt-optimization-loop-harness.js';
 
 describe('runPromptOptimizationLoop', () => {
-  test('resumes after a decided candidate and continues with the next round', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return roundId === 'round-0' ? 1 : 0;
-      };
-
-      const first = await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      assert.deepEqual(first.decisions.map((decision) => decision.decision), ['keep']);
-
-      const resumedMetaAgentRounds: string[] = [];
-      const resumedTaskRuns: string[] = [];
-      const resumed = await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 2,
-        baselineRuns: 1,
-        onTaskRun: (roundId, taskId) => resumedTaskRuns.push(`${roundId}:${taskId}`),
-        metaAgent: async (promptInput) => {
-          resumedMetaAgentRounds.push(promptInput.roundId);
-          return fakeMetaAgent()(promptInput);
-        },
-      });
-
-      assert.deepEqual(resumedMetaAgentRounds, ['round-1']);
-      assert.deepEqual(resumed.decisions.map((decision) => decision.decision), ['keep', 'discard']);
-      assert.equal(resumed.keptCount, 1);
-      assert.equal(resumed.stopReason, 'rounds_complete');
-      assert.equal(resumed.smoke.status, 'pass');
-      assert.ok(Math.abs(resumed.totalCostUsd - resumed.smoke.totalCostUsd) < 1e-9);
-      assert.ok(
-        resumedTaskRuns.every((call) => call.startsWith('round-1:')),
-        `unexpected resumed task runs: ${JSON.stringify(resumedTaskRuns)}`,
-      );
-
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      assert.equal(events.filter((event) => event.type === 'prompt_candidate_decided' && event.roundId === 'round-0').length, 1);
-      assert.equal(events.filter((event) => event.type === 'prompt_candidate_decided' && event.roundId === 'round-1').length, 1);
-    });
-  });
-
-  test('resumes after a committed candidate and finishes that round', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const commitIndex = events.findIndex((event) => event.type === 'prompt_candidate_committed' && event.roundId === 'round-0');
-      assert.ok(commitIndex > -1);
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${events.slice(0, commitIndex + 1).map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      const resumedMetaAgentRounds: string[] = [];
-      const resumed = await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-        metaAgent: async (promptInput) => {
-          resumedMetaAgentRounds.push(promptInput.roundId);
-          return fakeMetaAgent()(promptInput);
-        },
-      });
-
-      assert.deepEqual(resumedMetaAgentRounds, []);
-      assert.deepEqual(resumed.decisions.map((decision) => decision.decision), ['keep']);
-      assert.equal(resumed.smoke.status, 'pass');
-      const resumedEvents = await readFixedPromptWal(harness.resultsJsonlPath);
-      assert.equal(resumedEvents.filter((event) => event.type === 'prompt_candidate_committed' && event.roundId === 'round-0').length, 1);
-      assert.equal(resumedEvents.filter((event) => event.type === 'prompt_candidate_decided' && event.roundId === 'round-0').length, 1);
-    });
-  });
-
-  test('resumes after a committed candidate and discards back to the seed commit', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 0;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const committed = events.find((event): event is Extract<typeof event, { type: 'prompt_candidate_committed' }> =>
-        event.type === 'prompt_candidate_committed' && event.roundId === 'round-0');
-      assert.ok(committed);
-      const commitIndex = events.indexOf(committed);
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${events.slice(0, commitIndex + 1).map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-      await execFileAsync('git', ['reset', '--hard', committed.commitSha], { cwd: harness.repoDir });
-
-      const resumed = await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-
-      assert.deepEqual(resumed.decisions.map((decision) => decision.decision), ['discard']);
-      assert.equal(resumed.decisions[0]?.previousLastKeptCommitSha, harness.originalCommitSha);
-      assert.equal(resumed.lastKeptCommitSha, harness.originalCommitSha);
-      const head = (await execFileAsync('git', ['rev-parse', 'HEAD'], { cwd: harness.repoDir })).stdout.trim();
-      assert.equal(head, harness.originalCommitSha);
-      const resumedEvents = await readFixedPromptWal(harness.resultsJsonlPath);
-      const decision = resumedEvents.find((event): event is Extract<typeof event, { type: 'prompt_candidate_decided' }> =>
-        event.type === 'prompt_candidate_decided' && event.roundId === 'round-0');
-      assert.equal(decision?.lastKeptCommitSha, harness.originalCommitSha);
-    });
-  });
-
-  test('fails closed when replayed prompt decisions belong to a different resume fingerprint', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-        resumeFingerprint: 'fingerprint-old',
-      });
-
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-          resumeFingerprint: 'fingerprint-new',
-        }),
-        /RSI WAL replay identity mismatch/,
-      );
-    });
-  });
-
-  test('fails closed when replayed candidate task evidence has a stale prompt hash', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const staleEvents = events.map((event) => (
-        event.type === 'task_completed' && event.roundId === 'round-0' && event.taskId === 'hin-0'
-          ? { ...event, promptHash: 'sha256:stale' }
-          : event
-      ));
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${staleEvents.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-        }),
-        /RSI WAL replay prompt hash mismatch/,
-      );
-    });
-  });
-
-  test('fails closed when replayed task evidence has no prompt hash', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const missingHashEvents = events.map((event) => {
-        if (event.type !== 'task_completed' || event.roundId !== 'round-0' || event.taskId !== 'hin-0') return event;
-        const { promptHash: _promptHash, ...withoutPromptHash } = event;
-        return withoutPromptHash;
-      });
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${missingHashEvents.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-        }),
-        /RSI WAL replay prompt hash mismatch/,
-      );
-    });
-  });
-
-  test('fails closed when replayed baseline task evidence has a stale prompt hash', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const staleEvents = events.map((event) => (
-        event.type === 'task_completed' && event.roundId === 'baseline-0' && event.taskId === 'hin-0'
-          ? { ...event, promptHash: 'sha256:stale' }
-          : event
-      ));
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${staleEvents.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-        }),
-        /RSI WAL replay prompt hash mismatch/,
-      );
-    });
-  });
-
-  test('fails closed when replayed baseline task evidence has duplicate task ids', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const duplicate = events.find((event) =>
-        event.type === 'task_completed' && event.roundId === 'baseline-0' && event.taskId === 'hin-0');
-      assert.ok(duplicate);
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${[...events, duplicate].map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-        }),
-        /RSI WAL replay duplicate task event/,
-      );
-    });
-  });
-
-  test('fails closed when replaying task evidence without a resume fingerprint', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-        resumeFingerprint: null,
-      });
-
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-          resumeFingerprint: null,
-        }),
-        /RSI WAL replay requires a resume fingerprint/,
-      );
-    });
-  });
-
-  test('fails closed when task source changes under the same task ids', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-        resumeFingerprint: 'task-source-v1',
-      });
-
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks: heldInTasks.map((task) => ({ ...task, path: `${task.path}-changed` })),
-          heldOutTasks: heldOutTasks.map((task) => ({ ...task, path: `${task.path}-changed` })),
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-          resumeFingerprint: 'task-source-v2',
-        }),
-        /RSI WAL replay identity mismatch/,
-      );
-    });
-  });
-
-  test('fails closed instead of rerunning baseline when later WAL history exists', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const missingBaseline = events.filter((event) =>
-        !(event.type === 'task_completed' && event.roundId === 'baseline-0' && event.taskId === 'hin-0'));
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${missingBaseline.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      const rerunAttempts: string[] = [];
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-          onTaskRun: (roundId, taskId) => rerunAttempts.push(`${roundId}:${taskId}`),
-        }),
-        /RSI WAL replay missing required baseline held-in evidence for baseline-0/,
-      );
-      assert.deepEqual(rerunAttempts, []);
-    });
-  });
-
-  test('fails closed when a kept decision is missing held-out task evidence', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return taskId.startsWith('hin-') ? 1 : (index < 4 ? 1 : 0);
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const missingHeldOut = events.filter((event) =>
-        !(event.type === 'task_completed' && event.roundId === 'round-0' && event.taskId.startsWith('hout-')));
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${missingHeldOut.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      let nextRoundPrompted = false;
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-          metaAgent: async (promptInput) => {
-            if (promptInput.roundId === 'round-1') nextRoundPrompted = true;
-            return fakeMetaAgent()(promptInput);
-          },
-        }),
-        /RSI WAL replay missing required held-out task evidence for round-0/,
-      );
-      assert.equal(nextRoundPrompted, false);
-    });
-  });
-
-  test('fails closed when a held-out regression decision is missing held-out task evidence', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return roundId.startsWith('baseline-') && index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      const first = await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      assert.equal(first.decisions[0]?.reason, 'held_out_regressed');
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const missingHeldOut = events.filter((event) =>
-        !(event.type === 'task_completed' && event.roundId === 'round-0' && event.taskId.startsWith('hout-')));
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${missingHeldOut.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      let nextRoundPrompted = false;
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-          metaAgent: async (promptInput) => {
-            if (promptInput.roundId === 'round-1') nextRoundPrompted = true;
-            return fakeMetaAgent()(promptInput);
-          },
-        }),
-        /RSI WAL replay missing required held-out task evidence for round-0/,
-      );
-      assert.equal(nextRoundPrompted, false);
-    });
-  });
-
-  test('fails closed when a replayed decision is missing reward-hack scan evidence', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const withoutScan = events.map((event) => {
-        if (event.type !== 'prompt_candidate_decided' || event.roundId !== 'round-0') return event;
-        const { rewardHackScan: _rewardHackScan, ...withoutRewardHackScan } = event;
-        return withoutRewardHackScan;
-      });
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${withoutScan.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      let nextRoundPrompted = false;
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-          metaAgent: async (promptInput) => {
-            if (promptInput.roundId === 'round-1') nextRoundPrompted = true;
-            return fakeMetaAgent()(promptInput);
-          },
-        }),
-        /RSI WAL replay missing reward-hack scan evidence for round-0/,
-      );
-      assert.equal(nextRoundPrompted, false);
-    });
-  });
-
-  test('fails closed when a replayed decision disagrees with task evidence', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const tamperedDecision = events.map((event) => (
-        event.type === 'prompt_candidate_decided' && event.roundId === 'round-0'
-          ? { ...event, metrics: { tampered: true } }
-          : event
-      ));
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${tamperedDecision.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      let nextRoundPrompted = false;
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-          metaAgent: async (promptInput) => {
-            if (promptInput.roundId === 'round-1') nextRoundPrompted = true;
-            return fakeMetaAgent()(promptInput);
-          },
-        }),
-        /RSI WAL replay decision mismatch for round-0/,
-      );
-      assert.equal(nextRoundPrompted, false);
-    });
-  });
-
-  test('fails closed when a replayed decision is missing RSI attribution evidence', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const withoutAttribution = events.filter((event) =>
-        !(event.type === 'rsi_controller_attribution' && event.roundId === 'round-0'));
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${withoutAttribution.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      let nextRoundPrompted = false;
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-          metaAgent: async (promptInput) => {
-            if (promptInput.roundId === 'round-1') nextRoundPrompted = true;
-            return fakeMetaAgent()(promptInput);
-          },
-        }),
-        /RSI WAL replay missing post-decision RSI attribution evidence for round-0/,
-      );
-      assert.equal(nextRoundPrompted, false);
-    });
-  });
-
-  test('fails closed when RSI attribution appears before its decision', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const attributionIndex = events.findIndex((event) =>
-        event.type === 'rsi_controller_attribution' && event.roundId === 'round-0');
-      const decisionIndex = events.findIndex((event) =>
-        event.type === 'prompt_candidate_decided' && event.roundId === 'round-0');
-      assert.ok(attributionIndex > decisionIndex);
-      const attribution = events[attributionIndex]!;
-      const withoutAttribution = events.filter((_event, index) => index !== attributionIndex);
-      const decisionIndexAfterRemoval = withoutAttribution.findIndex((event) =>
-        event.type === 'prompt_candidate_decided' && event.roundId === 'round-0');
-      const attributionBeforeDecision = [
-        ...withoutAttribution.slice(0, decisionIndexAfterRemoval),
-        attribution,
-        ...withoutAttribution.slice(decisionIndexAfterRemoval),
-      ];
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${attributionBeforeDecision.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      let nextRoundPrompted = false;
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-          metaAgent: async (promptInput) => {
-            if (promptInput.roundId === 'round-1') nextRoundPrompted = true;
-            return fakeMetaAgent()(promptInput);
-          },
-        }),
-        /RSI WAL replay found RSI attribution before decision for round-0/,
-      );
-      assert.equal(nextRoundPrompted, false);
-    });
-  });
-
-  test('fails closed when RSI attribution appears after the next candidate', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return roundId === 'round-0' ? 1 : 0;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 2,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const attributionIndex = events.findIndex((event) =>
-        event.type === 'rsi_controller_attribution' && event.roundId === 'round-0');
-      const nextCandidateIndex = events.findIndex((event) =>
-        event.type === 'prompt_candidate_committed' && event.roundId === 'round-1');
-      assert.ok(attributionIndex > -1);
-      assert.ok(nextCandidateIndex > attributionIndex);
-      const attribution = events[attributionIndex]!;
-      const withoutAttribution = events.filter((_event, index) => index !== attributionIndex);
-      const nextCandidateIndexAfterRemoval = withoutAttribution.findIndex((event) =>
-        event.type === 'prompt_candidate_committed' && event.roundId === 'round-1');
-      const attributionAfterNextCandidate = [
-        ...withoutAttribution.slice(0, nextCandidateIndexAfterRemoval + 1),
-        attribution,
-        ...withoutAttribution.slice(nextCandidateIndexAfterRemoval + 1),
-      ];
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${attributionAfterNextCandidate.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-      const candidateCommitCountBefore = attributionAfterNextCandidate.filter((event) =>
-        event.type === 'prompt_candidate_committed').length;
-
-      let laterRoundPrompted = false;
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 3,
-          baselineRuns: 1,
-          metaAgent: async (promptInput) => {
-            if (promptInput.roundId === 'round-2') laterRoundPrompted = true;
-            return fakeMetaAgent()(promptInput);
-          },
-        }),
-        /RSI WAL replay missing post-decision RSI attribution evidence for round-0/,
-      );
-      assert.equal(laterRoundPrompted, false);
-      const eventsAfterResume = await readFixedPromptWal(harness.resultsJsonlPath);
-      assert.equal(
-        eventsAfterResume.filter((event) => event.type === 'prompt_candidate_committed').length,
-        candidateCommitCountBefore,
-      );
-      assert.equal(eventsAfterResume.some((event) =>
-        event.type === 'prompt_candidate_committed' && event.roundId === 'round-2'), false);
-    });
-  });
-
-  test('fails closed before prompting when replayed RSI attribution leaks held-out scope', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const tamperedAttribution = events.map((event) => (
-        event.type === 'rsi_controller_attribution' && event.roundId === 'round-0'
-          ? { ...event, predictedFixes: [{ taskId: 'hout-0', outcome: 'improved' }] }
-          : event
-      ));
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${tamperedAttribution.map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-
-      let nextRoundPrompted = false;
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 2,
-          baselineRuns: 1,
-          metaAgent: async (promptInput) => {
-            if (promptInput.roundId === 'round-1') nextRoundPrompted = true;
-            return fakeMetaAgent()(promptInput);
-          },
-        }),
-        /RSI WAL replay invalid RSI attribution evidence for round-0/,
-      );
-      assert.equal(nextRoundPrompted, false);
-    });
-  });
-
-  test('fails closed when prompt repo HEAD disagrees with WAL replay state', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const commitIndex = events.findIndex((event) =>
-        event.type === 'prompt_candidate_committed' && event.roundId === 'round-0');
-      assert.ok(commitIndex > -1);
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${events.slice(0, commitIndex + 1).map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-      await execFileAsync('git', ['reset', '--hard', harness.originalCommitSha], { cwd: harness.repoDir });
-
-      const taskRuns: string[] = [];
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 1,
-          baselineRuns: 1,
-          onTaskRun: (roundId, taskId) => taskRuns.push(`${roundId}:${taskId}`),
-        }),
-        /prompt repo HEAD does not match resumed RSI WAL state/,
-      );
-      assert.deepEqual(taskRuns, []);
-    });
-  });
-
-  test('fails closed before sweeping when a pending candidate task-set is stale', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 1;
-      };
-      const durationMsFor = (_roundId: string, taskId: string): number =>
-        taskId === 'hin-19' ? 200 : 10;
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        durationMsFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      const events = await readFixedPromptWal(harness.resultsJsonlPath);
-      const committed = events.find((event): event is Extract<typeof event, { type: 'prompt_candidate_committed' }> =>
-        event.type === 'prompt_candidate_committed' && event.roundId === 'round-0');
-      assert.ok(committed);
-      const commitIndex = events.indexOf(committed);
-      await writeFile(
-        harness.resultsJsonlPath,
-        `${events.slice(0, commitIndex + 1).map((event) => JSON.stringify(event)).join('\n')}\n`,
-        'utf8',
-      );
-      await execFileAsync('git', ['reset', '--hard', committed.commitSha], { cwd: harness.repoDir });
-
-      const taskRuns: string[] = [];
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          durationMsFor,
-          rounds: 1,
-          baselineRuns: 1,
-          maxStableTaskDurationMs: 100,
-          onTaskRun: (roundId, taskId) => taskRuns.push(`${roundId}:${taskId}`),
-        }),
-        /RSI WAL replay candidate task-set mismatch for round-0/,
-      );
-      assert.deepEqual(taskRuns, []);
-    });
-  });
-
-  test('fails closed before baseline when the WAL already belongs to another run', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return 0;
-      };
-
-      await runLoop(harness, {
-        runId: 'run-old',
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-
-      const taskRuns: string[] = [];
-      await assert.rejects(
-        runLoop(harness, {
-          runId: 'run-new',
-          heldInTasks,
-          heldOutTasks,
-          rewardFor,
-          rounds: 1,
-          baselineRuns: 1,
-          onTaskRun: (roundId, taskId) => taskRuns.push(`${roundId}:${taskId}`),
-        }),
-        /RSI WAL replay found events for a different runId/,
-      );
-      assert.deepEqual(taskRuns, []);
-    });
-  });
-
-  test('fails closed before baseline when prompt files are dirty', async () => {
-    await withHarness(async (harness) => {
-      const taskRuns: string[] = [];
-      await writeFile(harness.systemPromptPath, 'dirty prompt\n', 'utf8');
-
-      await assert.rejects(
-        runLoop(harness, {
-          heldInTasks: makeTasks('hin', 2),
-          heldOutTasks: makeTasks('hout', 1),
-          rewardFor: () => 1,
-          rounds: 1,
-          baselineRuns: 1,
-          onTaskRun: (roundId, taskId) => taskRuns.push(`${roundId}:${taskId}`),
-        }),
-        /prompt repo has uncommitted prompt file changes/,
-      );
-      assert.deepEqual(taskRuns, []);
-    });
-  });
-
-  test('rebuilds held-in TSV from WAL before prompting the next resumed round', async () => {
-    await withHarness(async (harness) => {
-      const heldInTasks = makeTasks('hin', 20);
-      const heldOutTasks = makeTasks('hout', 8);
-      const rewardFor = (roundId: string, taskId: string): number => {
-        const index = taskIndex(taskId);
-        if (taskId.startsWith('hout-')) return index < 4 ? 1 : 0;
-        if (roundId.startsWith('baseline-')) return index < 10 ? 1 : 0;
-        return roundId === 'round-0' ? 1 : 0;
-      };
-
-      await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 1,
-        baselineRuns: 1,
-      });
-      await rm(harness.heldInResultsTsvPath, { force: true });
-
-      let roundOneResultsTsv = '';
-      let diskHeldInTsvDuringPrompt = '';
-      const resumed = await runLoop(harness, {
-        heldInTasks,
-        heldOutTasks,
-        rewardFor,
-        rounds: 2,
-        baselineRuns: 1,
-        metaAgent: async (promptInput) => {
-          if (promptInput.roundId === 'round-1') {
-            roundOneResultsTsv = promptInput.resultsTsv;
-            diskHeldInTsvDuringPrompt = await readFile(harness.heldInResultsTsvPath, 'utf8');
-          }
-          return fakeMetaAgent()(promptInput);
-        },
-      });
-
-      assert.deepEqual(resumed.decisions.map((decision) => decision.decision), ['keep', 'discard']);
-      assert.match(roundOneResultsTsv, /^task_id\tstatus\tpassed\t/);
-      assert.match(roundOneResultsTsv, /hin-0\t/);
-      assert.doesNotMatch(roundOneResultsTsv, /hout-0\t/);
-      assert.match(diskHeldInTsvDuringPrompt, /hin-0\t/);
-      assert.doesNotMatch(diskHeldInTsvDuringPrompt, /hout-0\t/);
-    });
-  });
-
   test('keeps an improving candidate, discards a regressing one, and reports a passing smoke', async () => {
     await withHarness(async (harness) => {
       const heldInTasks = makeTasks('hin', 20);
@@ -1651,212 +563,3 @@ describe('runPromptOptimizationLoop', () => {
     });
   });
 });
-
-interface Harness {
-  repoDir: string;
-  controllerDir: string;
-  agentCwdPath: string;
-  programPath: string;
-  systemPromptPath: string;
-  resultsJsonlPath: string;
-  heldInResultsTsvPath: string;
-  heldOutResultsTsvPath: string;
-  eventsDir: string;
-  originalCommitSha: string;
-}
-
-interface RunLoopOptions {
-  runId?: string;
-  heldInTasks: readonly { id: string; path: string }[];
-  heldOutTasks: readonly { id: string; path: string }[];
-  rewardFor: (roundId: string, taskId: string) => number;
-  rounds: number;
-  baselineRuns: number;
-  costCeilingUsd?: number;
-  maxInfraFailureRate?: number;
-  heldOutResultsTsvPath?: string;
-  minStableHeldInTasks?: number;
-  maxStableTaskDurationMs?: number;
-  /** When it returns true, the runner emits a non-completed (unscored) cell for
-   * that task — used to exercise the baseline stability filter. */
-  shouldFail?: (roundId: string, taskId: string) => boolean;
-  /** Per-task baseline duration (ms); defaults to 10. Exercises the too-slow cap. */
-  durationMsFor?: (roundId: string, taskId: string) => number;
-  onTaskRun?: (roundId: string, taskId: string) => void;
-  metaAgent?: MetaAgent;
-  resumeFingerprint?: string | null;
-}
-
-async function runLoop(harness: Harness, options: RunLoopOptions) {
-  const nextId = idFactory();
-  let clock = 0;
-  const rewardHackVerifierPatternsByTaskId = Object.fromEntries(
-    options.heldInTasks.map((task) => [task.id, ['ZZZ_NO_VERIFIER_MATCH']]),
-  );
-  const input: PromptOptimizationLoopInput = {
-    runId: options.runId ?? 'run-1',
-    rounds: options.rounds,
-    baselineRuns: options.baselineRuns,
-    agentCwdPath: harness.agentCwdPath,
-    programPath: harness.programPath,
-    systemPromptPath: harness.systemPromptPath,
-    resultsJsonlPath: harness.resultsJsonlPath,
-    heldInResultsTsvPath: harness.heldInResultsTsvPath,
-    heldOutResultsTsvPath: options.heldOutResultsTsvPath ?? harness.heldOutResultsTsvPath,
-    heldInTasks: options.heldInTasks,
-    heldOutTasks: options.heldOutTasks,
-    config: CONFIG,
-    harborRunner: fakeHarborRunner(harness.eventsDir, options.rewardFor, options.shouldFail, options.durationMsFor, options.onTaskRun),
-    metaAgent: options.metaAgent ?? fakeMetaAgent(),
-    git: createCliPromptCandidateGit({ cwd: harness.repoDir, systemPromptPath: harness.systemPromptPath }),
-    rewardHackVerifierPatternsByTaskId,
-    ...(options.resumeFingerprint !== null ? { resumeFingerprint: options.resumeFingerprint ?? 'fingerprint-test' } : {}),
-    ...(options.costCeilingUsd !== undefined ? { costCeilingUsd: options.costCeilingUsd } : {}),
-    ...(options.maxInfraFailureRate !== undefined ? { maxInfraFailureRate: options.maxInfraFailureRate } : {}),
-    ...(options.minStableHeldInTasks !== undefined ? { minStableHeldInTasks: options.minStableHeldInTasks } : {}),
-    ...(options.maxStableTaskDurationMs !== undefined ? { maxStableTaskDurationMs: options.maxStableTaskDurationMs } : {}),
-    now: () => (clock += 1),
-    newId: nextId,
-  };
-  return runPromptOptimizationLoop(input);
-}
-
-/** A meta-agent that proposes a unique, valid prompt per round (no model). */
-function fakeMetaAgent(): MetaAgent {
-  return async (promptInput) => ({
-    systemPrompt: `candidate prompt ${promptInput.roundId}\n`,
-    summary: `tuned for ${promptInput.roundId}`,
-    candidateRationale: {
-      failurePattern: 'coverage_regression',
-      evidenceRefs: evidenceRefsFor(promptInput),
-      hypothesis: 'stable held-in coverage can improve with a clearer prompt',
-      targetedFix: 'make the success criteria explicit without adding task-specific answers',
-      predictedFixes: [],
-      riskTasks: [],
-    },
-  });
-}
-
-function evidenceRefsFor(promptInput: MetaAgentPromptInput): string[] {
-  const signal = promptInput.rsiAnalysis?.signals.find((item) => item.kind === 'coverage_regression')
-    ?? promptInput.rsiAnalysis?.signals[0];
-  return signal ? [signal.id] : [];
-}
-
-/** A Harbor runner that fabricates a completed, correctly-hashed cell per task
- * and writes a model-visible runtime-events file the digest/scan can read. */
-function fakeHarborRunner(
-  eventsDir: string,
-  rewardFor: (roundId: string, taskId: string) => number,
-  shouldFail?: (roundId: string, taskId: string) => boolean,
-  durationMsFor?: (roundId: string, taskId: string) => number,
-  onTaskRun?: (roundId: string, taskId: string) => void,
-): (input: HarborTaskRunInput) => Promise<HarborTaskRunOutput> {
-  return async ({ roundId, task, systemPrompt }) => {
-    onTaskRun?.(roundId, task.id);
-    const runtimeEventsPath = join(eventsDir, `${roundId}__${task.id}.jsonl`);
-    await writeFile(runtimeEventsPath, `${JSON.stringify(modelVisibleEvent())}\n`, 'utf8');
-    // A non-completed cell with a correct hash and real (non-zero) cost: scored
-    // is false, so the controller records it as an unscored task_completed — not
-    // a plumbing failure — which the stability filter drops.
-    const failed = shouldFail?.(roundId, task.id) ?? false;
-    return {
-      harbor: { reward: failed ? 0 : rewardFor(roundId, task.id) },
-      cell: {
-        schemaVersion: 1,
-        status: failed ? 'failed' : 'completed',
-        runtimeEventsPath,
-        promptHash: hashSystemPrompt(systemPrompt),
-        tokenSummary: tokenSummary({ input: 1, output: 2, reasoning: 0, total: 3, costUsd: COST_PER_TASK }),
-        toolSummary: {
-          providerVisibleToolCount: 1,
-          actualToolCalls: 1,
-          actualToolNames: ['Bash'],
-          actualToolCallCounts: { Bash: 1 },
-        },
-        steps: 1,
-        durationMs: durationMsFor?.(roundId, task.id) ?? 10,
-        startedAt: 0,
-        finishedAt: 10,
-        runtimeRefs: {
-          invocationId: `inv-${roundId}-${task.id}`,
-          sessionId: `session-${task.id}`,
-          runId: 'run-1',
-          turnId: `turn-${roundId}`,
-        },
-      },
-    };
-  };
-}
-
-function modelVisibleEvent(): unknown {
-  return {
-    id: 'call-1',
-    invocationId: 'inv-1',
-    runId: 'run-1',
-    sessionId: 'session-1',
-    turnId: 'turn-1',
-    ts: 1,
-    partial: false,
-    role: 'model',
-    author: 'agent',
-    content: { kind: 'function_call', id: 'call-1', name: 'Bash', args: { command: 'echo done' } },
-  };
-}
-
-function makeTasks(prefix: string, count: number): { id: string; path: string }[] {
-  return Array.from({ length: count }, (_unused, index) => ({
-    id: `${prefix}-${index}`,
-    path: `/tasks/${prefix}-${index}`,
-  }));
-}
-
-function taskIndex(taskId: string): number {
-  return Number(taskId.slice(taskId.lastIndexOf('-') + 1));
-}
-
-function idFactory(): () => string {
-  let counter = 0;
-  return () => `id-${(counter += 1)}`;
-}
-
-async function withHarness(fn: (harness: Harness) => Promise<void>): Promise<void> {
-  const root = await mkdtemp(join(tmpdir(), 'maka-prompt-loop-'));
-  try {
-    const repoDir = join(root, 'repo');
-    const controllerDir = join(root, 'controller');
-    const agentCwdPath = join(repoDir, 'agent-cwd');
-    const eventsDir = join(controllerDir, 'events');
-    await mkdir(repoDir, { recursive: true });
-    await mkdir(controllerDir, { recursive: true });
-    await mkdir(agentCwdPath, { recursive: true });
-    await mkdir(eventsDir, { recursive: true });
-
-    const programPath = join(repoDir, 'program.md');
-    const systemPromptPath = join(repoDir, 'system_prompt.md');
-    await writeFile(programPath, 'Improve the prompt conservatively.\n', 'utf8');
-    await writeFile(systemPromptPath, 'original prompt\n', 'utf8');
-
-    await execFileAsync('git', ['init'], { cwd: repoDir });
-    await execFileAsync('git', ['config', 'user.email', 'test@example.com'], { cwd: repoDir });
-    await execFileAsync('git', ['config', 'user.name', 'Test User'], { cwd: repoDir });
-    await execFileAsync('git', ['add', 'program.md', 'system_prompt.md'], { cwd: repoDir });
-    await execFileAsync('git', ['commit', '-m', 'initial'], { cwd: repoDir });
-    const originalCommitSha = (await execFileAsync('git', ['rev-parse', 'HEAD'], { cwd: repoDir })).stdout.trim();
-
-    await fn({
-      repoDir,
-      controllerDir,
-      agentCwdPath,
-      programPath,
-      systemPromptPath,
-      resultsJsonlPath: join(controllerDir, 'results.jsonl'),
-      heldInResultsTsvPath: join(controllerDir, 'held-in.tsv'),
-      heldOutResultsTsvPath: join(controllerDir, 'held-out.tsv'),
-      eventsDir,
-      originalCommitSha,
-    });
-  } finally {
-    await rm(root, { recursive: true, force: true });
-  }
-}
