+++ packages/headless/src/prompt-optimization-loop.ts
@@ -1,8 +1,4 @@
-import { execFile } from 'node:child_process';
 import { randomUUID } from 'node:crypto';
-import { realpath } from 'node:fs/promises';
-import { isAbsolute, relative, resolve } from 'node:path';
-import { isDeepStrictEqual, promisify } from 'node:util';
 import type { Config } from './contracts.js';
 import {
   appendFixedPromptWalEvent,
@@ -15,13 +11,11 @@ import {
   type FixedPromptTaskCompletedEvent,
   type FixedPromptTaskWalEvent,
   type HarborTaskRunner,
-  type PromptCandidateDecisionEvent,
   type PromptCandidateRewardHackScan,
   type RsiControllerAttributionEvent,
 } from './fixed-prompt-controller.js';
 import {
   extractTrajectoryDigest,
-  hashHeldInTaskSet,
   runPromptCandidateRound,
   scanRuntimeEventsForRewardHack,
   type MetaAgent,
@@ -50,12 +44,13 @@ import {
   type RsiPromptAttribution,
 } from './rsi-controller-attribution.js';
 import {
+  assertCandidateMatchesStableTaskSet,
+  assertPromptRepoMatchesReplayState,
+  assertReplayedDecisionMatchesResult,
   buildPromptOptimizationReplayPlan,
   replayPromptBaselinePartition,
   replayPromptDecisionRound,
 } from './prompt-optimization-replay.js';
-
-const execFileAsync = promisify(execFile);
 
 /**
  * Top-level driver for the RSI prompt-optimization loop (Issue #64).
@@ -648,85 +643,6 @@ export async function runPromptOptimizationLoop(
   };
 }
 
-async function assertPromptRepoMatchesReplayState(input: {
-  gitRootPath: string;
-  expectedHead: string;
-  programPath: string;
-  systemPromptGitPath: string;
-}): Promise<void> {
-  const head = await gitOutput(input.gitRootPath, 'rev-parse', 'HEAD');
-  if (head !== input.expectedHead) {
-    throw new Error(`prompt repo HEAD does not match resumed RSI WAL state: expected ${input.expectedHead}, got ${head}`);
-  }
-  const programGitPath = await toGitRelativePath(input.gitRootPath, input.programPath);
-  const promptGitPaths = [
-    programGitPath,
-    input.systemPromptGitPath,
-  ];
-  for (const path of [...new Set(promptGitPaths)]) {
-    if (!(await gitExitZero(input.gitRootPath, 'ls-files', '--error-unmatch', '--', path))) {
-      throw new Error(`prompt repo prompt file must be tracked before RSI run: ${path}`);
-    }
-  }
-  const [worktreeClean, indexClean] = await Promise.all([
-    gitExitZero(input.gitRootPath, 'diff', '--quiet', '--', ...promptGitPaths),
-    gitExitZero(input.gitRootPath, 'diff', '--cached', '--quiet', '--', ...promptGitPaths),
-  ]);
-  if (!worktreeClean || !indexClean) {
-    throw new Error(`prompt repo has uncommitted prompt file changes: ${promptGitPaths.join(', ')}`);
-  }
-}
-
-function assertReplayedDecisionMatchesResult(
-  decision: PromptCandidateDecisionEvent,
-  result: PromptAcceptanceResult,
-): void {
-  const replayedDecision = {
-    decision: result.decision,
-    reason: result.reason,
-    candidateCommitSha: result.candidateCommitSha,
-    previousLastKeptCommitSha: result.previousLastKeptCommitSha,
-    lastKeptCommitSha: result.lastKeptCommitSha,
-    previousHeldInReferencePassEligibleRate: result.previousHeldInReferencePassEligibleRate,
-    heldInReferencePassEligibleRate: result.heldInReferencePassEligibleRate,
-    originalCommitSha: result.originalCommitSha,
-    originalHeldOutPassEligibleRate: result.originalHeldOutPassEligibleRate,
-    heldInPassRateNoiseBand: result.heldInPassRateNoiseBand,
-    heldOutPassRateNoiseBand: result.heldOutPassRateNoiseBand,
-    rewardHackScan: result.rewardHackScan,
-    metrics: result.metrics,
-  };
-  const persistedDecision = {
-    decision: decision.decision,
-    reason: decision.reason,
-    candidateCommitSha: decision.candidateCommitSha,
-    previousLastKeptCommitSha: decision.previousLastKeptCommitSha,
-    lastKeptCommitSha: decision.lastKeptCommitSha,
-    previousHeldInReferencePassEligibleRate: decision.previousHeldInReferencePassEligibleRate,
-    heldInReferencePassEligibleRate: decision.heldInReferencePassEligibleRate,
-    originalCommitSha: decision.originalCommitSha,
-    originalHeldOutPassEligibleRate: decision.originalHeldOutPassEligibleRate,
-    heldInPassRateNoiseBand: decision.heldInPassRateNoiseBand,
-    heldOutPassRateNoiseBand: decision.heldOutPassRateNoiseBand,
-    rewardHackScan: decision.rewardHackScan,
-    metrics: decision.metrics,
-  };
-  if (!isDeepStrictEqual(persistedDecision, replayedDecision)) {
-    throw new Error(`RSI WAL replay decision mismatch for ${decision.roundId}`);
-  }
-}
-
-function assertCandidateMatchesStableTaskSet(
-  candidate: { roundId: string; heldInTaskIds: readonly string[]; heldInTaskSetHash: string },
-  stableHeldInTaskIds: readonly string[],
-): void {
-  const actualHash = hashHeldInTaskSet(candidate.heldInTaskIds);
-  const expectedHash = hashHeldInTaskSet(stableHeldInTaskIds);
-  if (candidate.heldInTaskSetHash !== actualHash || candidate.heldInTaskSetHash !== expectedHash) {
-    throw new Error(`RSI WAL replay candidate task-set mismatch for ${candidate.roundId}`);
-  }
-}
-
 function assertUniqueTaskIds(label: string, taskIds: readonly string[]): void {
   const seen = new Set<string>();
   const duplicates = new Set<string>();
@@ -745,30 +661,4 @@ function assertDisjointTaskIds(heldInTaskIds: readonly string[], heldOutTaskIds:
   if (overlap.length > 0) {
     throw new Error(`held-in and held-out tasks overlap: ${overlap.join(', ')}`);
   }
-}
-
-async function gitOutput(cwd: string, ...args: string[]): Promise<string> {
-  const { stdout } = await execFileAsync('git', args, { cwd, encoding: 'utf8' });
-  return stdout.trim();
-}
-
-async function gitExitZero(cwd: string, ...args: string[]): Promise<boolean> {
-  try {
-    await execFileAsync('git', args, { cwd });
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-async function toGitRelativePath(gitRootPath: string, filePath: string): Promise<string> {
-  const [rootPath, absolutePath] = await Promise.all([
-    realpath(gitRootPath),
-    realpath(isAbsolute(filePath) ? filePath : resolve(gitRootPath, filePath)),
-  ]);
-  const gitPath = relative(rootPath, absolutePath).split('\\').join('/');
-  if (gitPath === '' || gitPath === '..' || gitPath.startsWith('../')) {
-    throw new Error(`prompt repo prompt file must stay inside git root: ${filePath}`);
-  }
-  return gitPath;
 }
